# src/main/java/hu/davidder/webapp/core/base/pet/entity/PetXCategory.java
package hu.davidder.webapp.core.base.pet.entity;


import org.hibernate.annotations.Fetch;
import org.hibernate.annotations.FetchMode;

import com.fasterxml.jackson.annotation.JsonIgnore;

import hu.davidder.webapp.core.base.EntityBase;
import hu.davidder.webapp.core.base.category.entity.Category;
import jakarta.persistence.CascadeType;
import jakarta.persistence.Entity;
import jakarta.persistence.FetchType;
import jakarta.persistence.OneToOne;

@Entity
public class PetXCategory extends EntityBase {

    @JsonIgnore
    @OneToOne
    private Pet pet;

    @OneToOne(cascade = CascadeType.ALL, fetch = FetchType.LAZY, orphanRemoval = false)
    @Fetch(FetchMode.JOIN)
    private Category category;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public Pet getPet() {
        return pet;
    }

    public void setPet(Pet pet) {
        this.pet = pet;
    }

    public Category getCategory() {
        return category;
    }

    public void setCategory(Category category) {
        this.category = category;
    }
}